+++ app/components/application/index.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import Header from 'components/header';
-import TodoModal from 'components/todo_modal';
-import SigninModal from 'components/signin_modal';
-import SignupModal from 'components/signup_modal';
+import Modals from 'components/modals';
 import Footer from 'components/footer';
 import styles from './styles';
 
@@ -13,11 +11,9 @@ export default class Application extends React.Component {
         <main className={ styles.wrapper }>
           <Header/>
           { this.props.children }
-          <TodoModal/>
-          <SigninModal/>
-          <SignupModal/>
         </main>
         <Footer/>
+        <Modals/>
       </div>
     );
   }