+++ app/components/modals/tests/index.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import Modals from 'components/modals';
+
+describe('Modals', () => {
+  const modalsComponent = mount(<Modals/>);
+
+  it('renders wrapper', () => {
+    expect(modalsComponent.find('div').length).toEqual(1);
+  });
+});