+++ app/stores/application.js
@@ -1,0 +1,28 @@
+import Alt from 'alt_flux';
+import { createStore } from 'alt-utils/lib/decorators';
+import ApplicationActions from 'actions/application';
+
+@createStore(Alt)
+export default class ApplicationStore {
+  constructor() {
+    this.isModalOpen = false;
+    this.modalName = '';
+    this.modalOptions = {};
+
+    this.bindListeners({
+      openModal: ApplicationActions.OPEN_MODAL,
+      closeModal: ApplicationActions.CLOSE_MODAL
+    });
+  }
+
+  openModal({ name, ...rest }) {
+    this.isModalOpen = true;
+    this.modalName = name;
+    this.modalOptions = { ...rest };
+  }
+
+  closeModal() {
+    this.isModalOpen = false;
+    this.modalName = '';
+  }
+}