+++ app/stores/todo.js
@@ -1,12 +1,10 @@
 import Alt from 'alt_flux';
 import { createStore } from 'alt-utils/lib/decorators';
-import { mixin } from 'core-decorators';
-import { ModalStoreMixin } from 'mixins/modal/store';
 import TodoActions from 'actions/todo';
+import ApplicationActions from 'actions/application';
 
 @createStore(Alt)
-@mixin(ModalStoreMixin)
-export default class TodosStore {
+export default class TodoStore {
   defaultProps = {
     name: '',
     isComplete: false
@@ -16,18 +14,16 @@ export default class TodosStore {
     this.todo = Object.assign({}, this.defaultProps);
 
     this.bindListeners({
-      show: TodoActions.SHOW,
-      hide: TodoActions.HIDE,
       setName: TodoActions.SET_NAME,
-      reset: TodoActions.RESET
+      reset: ApplicationActions.CLOSE_MODAL
     });
-  }
-
-  reset() {
-    this.todo = Object.assign({}, this.defaultProps);
   }
 
   setName(name) {
     this.todo.name = name;
+  }
+
+  reset() {
+    this.todo = Object.assign({}, this.defaultProps);
   }
 }