+++ config/karma.js
@@ -57,17 +57,7 @@ module.exports = (karmaConfig) => {
       }
     },
     webpackMiddleware: {
-      quiet: false,
-      noInfo: false,
-      stats: {
-        assets: false,
-        colors: true,
-        version: false,
-        hash: false,
-        timings: false,
-        chunks: false,
-        chunkModules: false
-      }
+      quiet: true
     }
   });
 };